Navigation Refactoring
Navigation is now handled using NavigationElements and
IntentNavigationElements and respective sublclasses.

This makes it simpler and much quicker to implement new functionality
for the app. Simply make a View in the Fragment_Navigation layout, wire
up the view inside of NavigationFragment using the id you assigned the
View, and create a class that has the right behavior on being clicked.