Triangle Rendering Fixes
Sobel Filter was the culprit, but StickyPointMaker was also O(n^2), so
that was removed.

Point population looks like this now:
PointMaker -> PixelPointMaker -> EdgePointMaker -> RadiusPointMaker ->
RandomPointMaker

Each of these extends the one before it, with the top level (PointMaker)
being an interface. The names should make it obvious how they populate
point, but I'll describe it anyway.

PixelPointMaker is abstract, so it doesn't actually populate anything;
it just sets up some basic variables which are used by all of the later
classes.

EdgePointMaker applies the sobel filter to the image, and populates
uniformly distributed points up to a limit from the edges that are
found.

RadiusPointMaker populates points based off of the points populated from
EdgePointMaker, it's the same philosophy as StickyPointMaker in that we
want much more information near the edges than far away from them so we
don't waste points. The difference here is that for each point from
EdgeMaker, points are populated using a Normal distribution within a
certain Radius (hence the name RadiusPointMaker).

RandomPointMaker extends these only for convenience, it doesn't take
into account any of the points that have been populated previously, only
in so far as to ensure that if there are already too many points, it
won't add any. The points are added as a final step to ensure that even
if there were no edges detected in the image (perhaps too blurry?) we
will still have at least a certain amount of points on the final image.