Image Options Implemented
You can now choose to add a single point at a time, points by dragging,
or to auto-triangulate an image. Also, you now have the option to change
the way the triangles are drawn on top of the image- you can see a wire
frame, the triangles on top of the image, the triangles drawn with
transparency on top of the image, or a combination of the previous
options. Also, the 'default images' are now removed from the app.