Fix
EditText was crashing with no text in it because it was setting
width/height to 1. This caused an issue in the Triangulation. Minimum
dimensions are now 2px by 2px, and this is verified before trying to
make a Gradient.

Also, fixed behavior when selecting only one color. It no longer
crashes, but instead makes a gradient from whatever color was selected
to black. This could have better behavior, but it works for the moment.